build: when debugging, do not remove build directories
Keep build directories in debug mode.

+++ go.sh
@@ -111,6 +111,6 @@ for t in ${TARGETS}; do
 	fi
 
 	popd
-	rm -rf  build_${t}
+	test -n ${DEBUG_BUILD} || rm -rf  build_${t}
 
 done
